+++ src/Models/Category/Category.php
@@ -48,7 +48,7 @@ class Category extends Model
 
         // If this field is not storing it's data on the category_field_data table we
         // will join the separate data table with a unique orderby_field_name alias
-        if ($field->legacy_field_data == 'n' || $field->legacy_field_data === false) {
+        if ($field->legacy_field_data === 'n' || $field->legacy_field_data === false) {
             $alias = "orderby_{$field->field_name}";
             $column = "$alias.$column";
             $this->scopeJoinFieldDataTable($query, $field, $alias);