+++ CHANGELOG.md
@@ -34,6 +34,7 @@
 - Issue with registering a fluid field that has an empty string in it's list of fields
 - Getting custom field data on members and categories through GraphQL
 - Bug with Fieldtypes providing certain GraphQL types
+- Bug checking if a field stored data on the legacy data table
 
 ### Changed